updates to content, and naming scheme

+++ content/contest/template.cpp
@@ -7,10 +7,10 @@ typedef vector<int> vi;
 
 const ll oo = 0x3f3f3f3f3f3f3f3fLL;
 
-#define FOR(i, a, b) for(ll i = (a); i < int(b); i++)
-#define FORD(i, a, b) for(ll i = (b)-1; i >= int(a); i--)
+#define FOR(i, a, b) for(int i = (a); i < int(b); i++)
+#define FORD(i, a, b) for(int i = (b)-1; i >= int(a); i--)
 #define has(c, e) ((c).find(e) != (c).end())
-#define sz(c) ll((c).size())
+#define sz(c) int((c).size())
 #define all(c) c.begin(), c.end()
 
 int main() {
